+++ src/base/b-virtual-scroll/modules/chunk-render.ts
@@ -17,7 +17,7 @@ import type bVirtualScroll from 'base/b-virtual-scroll/b-virtual-scroll';
 import type ComponentRender from 'base/b-virtual-scroll/modules/component-render';
 import type ChunkRequest from 'base/b-virtual-scroll/modules/chunk-request';
 
-import type { RenderItem } from 'base/b-virtual-scroll/interface';
+import type { RenderItem, VirtualItemEl } from 'base/b-virtual-scroll/interface';
 
 export const
 	$$ = symbolGenerator();
@@ -289,8 +289,14 @@ export default class ChunkRender extends Friend {
 
 			item.node = node;
 
+			const itemsData = {
+				current: item.data,
+				prev: items[i - 1]?.data,
+				next: items[i + 1]?.data
+			};
+
 			if (!Object.isFunction(node[$$.inView])) {
-				this.wrapInView(item);
+				this.wrapInView(item, itemsData);
 			}
 		}
 
@@ -300,8 +306,9 @@ export default class ChunkRender extends Friend {
 	/**
 	 * Wraps the specified item node with the `in-view` directive
 	 * @param item
+	 * @param itemData
 	 */
-	protected wrapInView(item: RenderItem): void {
+	protected wrapInView(item: RenderItem, itemData: VirtualItemEl): void {
 		const
 			{ctx} = this,
 			{node} = item;
@@ -311,7 +318,7 @@ export default class ChunkRender extends Friend {
 		}
 
 		const
-			label = `${this.asyncGroup}:${this.asyncInViewPrefix}${ctx.getItemKey(item.data, item.index)}`;
+			label = `${this.asyncGroup}:${this.asyncInViewPrefix}${ctx.getItemKey(itemData, item.index)}`;
 
 		if (!node) {
 			return;